requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
pydantic-settings>=2.0
google-cloud-spanner>=3.40.0
google-auth>=2.0
httpx>=0.27.0
google-cloud-storage>=2.14.0
google-api-python-client>=2.100.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=0.6.23
